+++ offchain/package.json
@@ -50,7 +50,7 @@
     "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/preset-env": "^7.27.2",
     "@babel/preset-typescript": "^7.27.1",
-    "@blaze-cardano/emulator": "^0.3.25",
+    "@blaze-cardano/emulator": "link:@blaze-cardano/emulator",
     "@blaze-cardano/blueprint": "^0.6.0",
     "@sundaeswap/babel-preset": "^2.0.15",
     "@sundaeswap/eslint-config": "^2.0.15",
@@ -66,7 +66,7 @@
   "dependencies": {
     "@blaze-cardano/data": "^0.6.0",
     "@blaze-cardano/query": "link:@blaze-cardano/query",
-    "@blaze-cardano/sdk": "^0.2.30",
+    "@blaze-cardano/sdk": "link:@blaze-cardano/sdk",
     "@blaze-cardano/uplc": "^0.3.1",
     "@inquirer/prompts": "^7.5.3",
     "cbor": "^10.0.3",
